+++ libsolidity/analysis/ControlFlowGraph.cpp
@@ -34,15 +34,34 @@ bool CFG::constructFlow(ASTNode const& _astRoot)
 
 bool CFG::visit(FunctionDefinition const& _function)
 {
-	if (_function.isImplemented())
-		m_functionControlFlow[&_function] = ControlFlowBuilder::createFunctionFlow(m_nodeContainer, _function);
+	if (_function.isImplemented() && _function.isFree())
+		m_functionControlFlow[pair(nullptr, &_function)] = ControlFlowBuilder::createFunctionFlow(m_nodeContainer, _function, nullptr);
 	return false;
 }
 
-FunctionFlow const& CFG::functionFlow(FunctionDefinition const& _function) const
+bool CFG::visit(ContractDefinition const& _contract)
 {
-	solAssert(m_functionControlFlow.count(&_function), "");
-	return *m_functionControlFlow.find(&_function)->second;
+	set<CallableDeclaration const*> overridden;
+
+	for (ContractDefinition const* contract: _contract.annotation().linearizedBaseContracts)
+		for (FunctionDefinition const* function: contract->definedFunctions())
+		{
+			if (overridden.count(function))
+				continue;
+
+			for (auto const* function: function->annotation().baseFunctions)
+				overridden.emplace(function);
+
+			m_functionControlFlow[pair(&_contract, function)] =
+				ControlFlowBuilder::createFunctionFlow(m_nodeContainer, *function, &_contract);
+		}
+
+	return true;
+}
+
+FunctionFlow const& CFG::functionFlow(ContractDefinition const& _contract, FunctionDefinition const& _function) const
+{
+	return *m_functionControlFlow.at(pair(&_contract, &_function));
 }
 
 CFGNode* CFG::NodeContainer::newNode()